Renamed Iowa bar component to Brand Bar .
