+++ src/components/brand-bar/brand-bar.scss
