+++ src/components/brand-bar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './BrandBar.vue'
