+++ src/components/index.ts
@@ -1,4 +1,4 @@
 /* eslint-disable import/prefer-default-export */
 
 export { default as UidsLogo } from './logo/Logo.vue';
-export { default as UidsIowaBar } from './iowa-bar/'
+export { default as UidsIowaBar } from './brand-bar/'
