--- src/components/iowa-bar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './IowaBar.vue'
